fix: extract shared schema helpers, add JSDoc, improve clarity (#84) (#89)
Extract getTopologicalOrder, getRelationshipInfo, and sanitizeTableName
into shared schema-helpers.ts — removes identical duplicates from both
mysql and postgres schema-introspector files. Add JSDoc to HookContext
interface fields, add explicit return to auth(), and improve store
unload warning messages with remediation hints.
